create QgsRectangleItemBase as base of QgsPictureItem
