+++ python/kadasgui/auto_generated/mapitems/kadaspictureitem.sip.in
@@ -10,7 +10,7 @@
 
 
 
-class KadasPictureItem : KadasMapItem
+class KadasPictureItem : KadasRectangleItemBase
 {
 %Docstring(signature="appended")
 *************************************************************************
@@ -33,76 +33,18 @@ the Free Software Foundation; either version 2 of the License, or     *
 
     const QString &filePath() const;
     void setFilePath( const QString &filePath );
-    bool frameVisible() const;
-    void setFrameVisible( bool frame );
-    bool positionLocked() const;
-    void setPositionLocked( bool locked );
 
     virtual QString itemName() const;
 
     virtual QImage symbolImage() const;
 
-    virtual KadasItemRect boundingBox() const;
-
-    virtual Margin margin() const;
-
-    virtual QList<KadasMapItem::Node> nodes( const QgsMapSettings &settings ) const;
-
-    virtual bool intersects( const KadasMapRect &rect, const QgsMapSettings &settings, bool contains = false ) const;
-
-    virtual void render( QgsRenderContext &context ) const;
-
-
-    virtual bool startPart( const KadasMapPos &firstPoint, const QgsMapSettings &mapSettings );
-
-    virtual bool startPart( const AttribValues &values, const QgsMapSettings &mapSettings );
-
-    virtual void setCurrentPoint( const KadasMapPos &p, const QgsMapSettings &mapSettings );
-
-    virtual void setCurrentAttributes( const AttribValues &values, const QgsMapSettings &mapSettings );
-
-    virtual bool continuePart( const QgsMapSettings &mapSettings );
-
-    virtual void endPart();
-
-
-    virtual AttribDefs drawAttribs() const;
-
-    virtual AttribValues drawAttribsFromPosition( const KadasMapPos &pos, const QgsMapSettings &mapSettings ) const;
-
-    virtual KadasMapPos positionFromDrawAttribs( const AttribValues &values, const QgsMapSettings &mapSettings ) const;
-
-
-    virtual EditContext getEditContext( const KadasMapPos &pos, const QgsMapSettings &mapSettings ) const;
-
-    virtual void edit( const EditContext &context, const KadasMapPos &newPoint, const QgsMapSettings &mapSettings );
-
-    virtual void edit( const EditContext &context, const AttribValues &values, const QgsMapSettings &mapSettings );
-
-    virtual void populateContextMenu( QMenu *menu, const EditContext &context, const KadasMapPos &clickPos, const QgsMapSettings &mapSettings );
 
     virtual void onDoubleClick( const QgsMapSettings &mapSettings );
 
 
-    virtual AttribValues editAttribsFromPosition( const EditContext &context, const KadasMapPos &pos, const QgsMapSettings &mapSettings ) const;
 
-    virtual KadasMapPos positionFromEditAttribs( const EditContext &context, const AttribValues &values, const QgsMapSettings &mapSettings ) const;
-
-
-    virtual KadasItemPos position() const;
-    virtual void setPosition( const KadasItemPos &pos );
-
-
-
-    struct State : KadasMapItem::State
+    struct State : KadasRectangleItemBase::State
     {
-      KadasItemPos pos;
-      QList<KadasItemPos> footprint;
-      KadasItemPos cameraTarget;
-      double angle;
-      double offsetX;
-      double offsetY;
-      QSize size;
       virtual void assign( const KadasMapItem::State *other );
       virtual State *clone() const /Factory/;
       virtual QJsonObject serialize() const;
@@ -110,13 +52,16 @@ the Free Software Foundation; either version 2 of the License, or     *
       virtual bool deserialize( const QJsonObject &json );
 
     };
-    const State *constState() const;
     virtual void setState( const KadasMapItem::State *state );
 
 
   protected:
     virtual KadasMapItem *_clone() const;
     virtual State *createEmptyState() const;
+    virtual void renderPrivate( QgsRenderContext &context, double dpiScale, double offsetX, double offsetY, double width, double height  ) const;
+
+    virtual void editPrivate( const KadasMapPos &newPoint, const QgsMapSettings &mapSettings );
+
 
 };
