+++ python/kadasgui/auto_generated/mapitems/kadasrectangleitembase.sip.in
@@ -1,0 +1,118 @@
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * kadas/gui/mapitems/kadasrectangleitembase.h                          *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
+
+
+
+
+
+class KadasRectangleItemBase : KadasMapItem
+{
+%Docstring(signature="appended")
+*************************************************************************
+
+This program is free software; you can redistribute it and/or modify  *
+it under the terms of the GNU General Public License as published by  *
+the Free Software Foundation; either version 2 of the License, or     *
+(at your option) any later version.                                   *
+
+**************************************************************************
+%End
+
+%TypeHeaderCode
+#include "kadas/gui/mapitems/kadasrectangleitembase.h"
+%End
+  public:
+    KadasRectangleItemBase( const QgsCoordinateReferenceSystem &crs );
+    virtual ~KadasRectangleItemBase();
+
+    bool frameVisible() const;
+    void setFrameVisible( bool frame );
+    bool positionLocked() const;
+    void setPositionLocked( bool locked );
+
+    virtual KadasItemRect boundingBox() const;
+
+    virtual Margin margin() const;
+
+    virtual QList<KadasMapItem::Node> nodes( const QgsMapSettings &settings ) const;
+
+    virtual bool intersects( const KadasMapRect &rect, const QgsMapSettings &settings, bool contains = false ) const;
+
+    virtual void render( QgsRenderContext &context ) const;
+
+
+    virtual bool startPart( const KadasMapPos &firstPoint, const QgsMapSettings &mapSettings );
+
+    virtual bool startPart( const AttribValues &values, const QgsMapSettings &mapSettings );
+
+    virtual void setCurrentPoint( const KadasMapPos &p, const QgsMapSettings &mapSettings );
+
+    virtual void setCurrentAttributes( const AttribValues &values, const QgsMapSettings &mapSettings );
+
+    virtual bool continuePart( const QgsMapSettings &mapSettings );
+
+    virtual void endPart();
+
+
+    virtual AttribDefs drawAttribs() const;
+
+    virtual AttribValues drawAttribsFromPosition( const KadasMapPos &pos, const QgsMapSettings &mapSettings ) const;
+
+    virtual KadasMapPos positionFromDrawAttribs( const AttribValues &values, const QgsMapSettings &mapSettings ) const;
+
+
+    virtual EditContext getEditContext( const KadasMapPos &pos, const QgsMapSettings &mapSettings ) const;
+
+    virtual void edit( const EditContext &context, const KadasMapPos &newPoint, const QgsMapSettings &mapSettings );
+
+    virtual void edit( const EditContext &context, const AttribValues &values, const QgsMapSettings &mapSettings );
+
+    virtual void populateContextMenu( QMenu *menu, const EditContext &context, const KadasMapPos &clickPos, const QgsMapSettings &mapSettings );
+
+
+    virtual AttribValues editAttribsFromPosition( const EditContext &context, const KadasMapPos &pos, const QgsMapSettings &mapSettings ) const;
+
+    virtual KadasMapPos positionFromEditAttribs( const EditContext &context, const AttribValues &values, const QgsMapSettings &mapSettings ) const;
+
+
+    virtual KadasItemPos position() const;
+    virtual void setPosition( const KadasItemPos &pos );
+
+
+
+    struct State : KadasMapItem::State
+    {
+      KadasItemPos mPos;
+      QList<KadasItemPos> mFootprint;
+      KadasItemPos mAnchorPoint;
+      double mAngle;
+      double mOffsetX;
+      double mOffsetY;
+      QSize mSize;
+      virtual void assign( const KadasMapItem::State *other );
+      virtual State *clone() const /Factory/;
+      virtual QJsonObject serialize() const;
+      virtual bool deserialize( const QJsonObject &json );
+    };
+    const State *constState() const;
+    virtual void setState( const KadasMapItem::State *state );
+
+  protected:
+    virtual State *createEmptyState() const;
+    virtual void renderPrivate( QgsRenderContext &context, double dpiScale, double offsetX, double offsetY, double width, double height ) const = 0;
+    virtual void editPrivate( const KadasMapPos &newPoint, const QgsMapSettings &mapSettings ) = 0;
+
+};
+
+/************************************************************************
+ * This file has been generated automatically from                      *
+ *                                                                      *
+ * kadas/gui/mapitems/kadasrectangleitembase.h                          *
+ *                                                                      *
+ * Do not edit manually ! Edit header and run scripts/sipify.pl again   *
+ ************************************************************************/
